Accidental 4-5 hour session.. lol its 230
Created enemy prefab, who is a bike tire with a nail and spearmint gum
Created ai that walks towards the player, and is supposed to jump if the player is higher than him, but this is bugged and it always jumps.
Created spawning system that needs tweaking, but it can spawn enemies. As of now enemy deaths cannot happen, and are not tracked. This needs to happen and update a variable in the Grid component for enemy spawning for it to work properly.
Attempted to work on a very basic collision and attack system, but it was breaking the game so I commented it out.

Cheers